5MLRE/Day4-TP1/requirements.txt
pandas~=1.5.2
colorama~=0.4.6
scikit-surprise~=1.1.3
numpy~=1.22.3